+++ example/package.json
@@ -14,10 +14,17 @@
   },
   "dependencies": {
     "@react-native-community/slider": "^4.5.2",
+    "@react-navigation/bottom-tabs": "^6.5.20",
+    "@react-navigation/native": "^6.1.17",
+    "@react-navigation/native-stack": "^6.9.26",
     "@theoplayer/react-native-ui": "^0.8.0",
+    "install": "^0.13.0",
+    "npm": "^10.5.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-native": "npm:react-native-tvos@^0.74.1-0",
+    "react-native-safe-area-context": "^4.9.0",
+    "react-native-screens": "^3.29.0",
     "react-native-status-bar-height": "^2.6.0",
     "react-native-svg": "^13.14.1",
     "react-native-web": "^0.19.12",
