+++ example/src/screens/ConfigScreen.tsx
@@ -1,0 +1,12 @@
+import * as React from 'react';
+import { View, Text } from 'react-native';
+
+function ConfigScreen() {
+  return (
+    <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+      <Text>Config Screen</Text>
+    </View>
+  );
+}
+
+export default ConfigScreen;
